% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPWRFisher.R
\name{fitPWRFisher}
\alias{fitPWRFisher}
\title{fitPWRFisher is used to fit a PWR model.}
\usage{
fitPWRFisher(X, Y, K, p = 3)
}
\arguments{
\item{X}{Numeric vector of length \emph{m} representing the covariates/inputs
\eqn{x_{1},\dots,x_{m}}.}

\item{Y}{Numeric vector of length \emph{m} representing the observed
response/output \eqn{y_{1},\dots,y_{m}}.}

\item{K}{The number of regimes (PWR components).}

\item{p}{Optional. The order of the polynomial regression. By default, \code{p} is
set at 3.}
}
\value{
fitPWRFisher returns an object of class \link{ModelPWR}.
}
\description{
fitPWRFisher fits a Piecewise Regression (PWR) model to a regime changing time-series (or a longitidunal data). The estimation method is performed by an optimized dynamic programming algorithm.
}
\details{
fitPWRFisher function implements the dynamic programming algorithm for piece-wise regression model.
This function starts with the calculation of the "cost matrix" then it
estimates the transition points given \code{K} the number of regimes thanks to
the method \code{computeDynamicProgram} (method of the class
\link{ParamPWR}).
}
\seealso{
\link{ModelPWR}, \link{ParamPWR}, \link{StatPWR}
}
