% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MData.R
\docType{class}
\name{MData-class}
\alias{MData-class}
\alias{MData}
\title{A Reference Class which represents multivariate data.}
\description{
MData is a reference class which represents multivariate objects. The data
can be ordered by time (multivariate time series). In the last case, the
field \code{X} represents the time.
}
\section{Fields}{

\describe{
\item{\code{X}}{Numeric vector of length \emph{m}.}

\item{\code{Y}}{Matrix of size \eqn{(m, d)} representing a \eqn{d} dimension
function of \code{X} observed at points \eqn{1,\dots,m}.}
}}


