% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkStochastic.R
\name{mkStochastic}
\alias{mkStochastic}
\title{mkStochastic ensures that it is a stochastic vector, matrix or array.}
\usage{
mkStochastic(M)
}
\arguments{
\item{M}{A vector, matrix or array to transform.}
}
\value{
A vector, matrix or array for which the sum over the last dimension
is 1.
}
\description{
mkStochastic ensures that it is a stochastic vector, matrix or array.
}
\details{
mkStochastic ensures that the giving argument is a stochastic
vector, matrix or array, i.e., that the sum over the last dimension is 1.
}
