% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMRHLP.R
\docType{class}
\name{ModelMRHLP-class}
\alias{ModelMRHLP-class}
\alias{ModelMRHLP}
\title{A Reference Class which represents a fitted MRHLP model.}
\description{
ModelMRHLP represents an estimated MRHLP model.
}
\section{Fields}{

\describe{
\item{\code{param}}{A \link{ParamMRHLP} object. It contains the
estimated values of the parameters.}

\item{\code{stat}}{A \link{StatMRHLP} object. It contains all the
statistics associated to the MRHLP model.}
}}

\section{Methods}{

\describe{
\item{\code{plot(what = c("regressors", "estimatedsignal"), ...)}}{Plot method.
\describe{
  \item{\code{what}}{The type of graph requested:
    \itemize{
      \item \code{"regressors" = } Polynomial regression components
        (fields \code{polynomials} and \code{pi_ik} of class
        \link{StatMRHLP}).
      \item \code{"estimatedsignal" = } Estimated signal (fields
        \code{Ex} and \code{klas} of class \link{StatMRHLP}).
    }
  }
  \item{\code{\dots}}{Other graphics parameters.}
}
By default, all the above graphs are produced.}

\item{\code{summary(digits = getOption("digits"))}}{Summary method.
\describe{
  \item{\code{digits}}{The number of significant digits to use when
    printing.}
}}
}}

\seealso{
\link{ParamMRHLP}, \link{StatMRHLP}
}
