% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-multivrealdataset.R
\docType{data}
\name{multivrealdataset}
\alias{multivrealdataset}
\title{Time series representing the three acceleration components recorded over time
during the activity of a given person.}
\format{A data frame with 2253 rows and 4 columns:
\describe{
\item{x}{The covariate variable which is the time in that case.}
\item{y1}{X axis component of the acceleration.}
\item{y2}{Y axis component of the acceleration.}
\item{y3}{Z axis component of the acceleration.}
}}
\usage{
multivrealdataset
}
\description{
This dataset provided for illustration only, represents the three
acceleration components recorded over time during the activity of a given
person. These data consist therefore in multidimensional time series with
several regime changes over time, each regime is associated with an activity.
}
\keyword{datasets}
