% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectMHMMR.R
\name{selectMHMMR}
\alias{selectMHMMR}
\title{selectMHMMR implements the model selection procedure.}
\usage{
selectMHMMR(X, Y, Kmin = 1, Kmax = 10, pmin = 0, pmax = 4,
  criterion = c("BIC", "AIC"))
}
\arguments{
\item{X}{Numeric vector of length \emph{m} representing the covariates/inputs
\eqn{x_{1},\dots,x_{m}}.}

\item{Y}{Matrix of size \eqn{(m, d)} representing a \eqn{d} dimension
time series observed at points \eqn{1,\dots,m}.}

\item{Kmin}{The minimum number of regimes (c components).}

\item{Kmax}{The maximum number of regimes (MHMMR components).}

\item{pmin}{The minimum order of the polynomial regression.}

\item{pmax}{The maximum order of the polynomial regression.}

\item{criterion}{The criterion used to select the MHMMR model ("BIC", "AIC").}
}
\value{
selectMHMMR returns an object of class \link{ModelMHMMR}
representing the selected MHMMR model according to the chosen \code{criterion}.
}
\description{
selectMHMMR implements the model selection procedure.
}
\details{
selectMHMMR implements the model selection. This function runs every
MHMMR model by varying the number of regimes \code{K} from \code{Kmin} to \code{Kmax} and
the order of the polynomial regression \code{p} from \code{pmin} to \code{pmax}. The model
having the highest value of the chosen selection criterion is then
selected.
}
\seealso{
\link{ModelMHMMR}
}
