% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectRHLP.R
\name{selectRHLP}
\alias{selectRHLP}
\title{selectRHLP implements the model selection procedure.}
\usage{
selectRHLP(X, Y, Kmin = 1, Kmax = 10, pmin = 0, pmax = 4,
  criterion = c("BIC", "AIC"))
}
\arguments{
\item{X}{Numeric vector of length \emph{m} representing the covariates/inputs
\eqn{x_{1},\dots,x_{m}}.}

\item{Y}{Numeric vector of length \emph{m} representing the observed
response/output \eqn{y_{1},\dots,y_{m}}.}

\item{Kmin}{The minimum number of regimes (RHLP components).}

\item{Kmax}{The maximum number of regimes (RHLP components).}

\item{pmin}{The minimum order of the polynomial regression.}

\item{pmax}{The maximum order of the polynomial regression.}

\item{criterion}{The criterion used to select the RHLP model ("BIC", "AIC").}
}
\value{
selectRHLP returns an object of class \link{ModelRHLP}
representing the selected RHLP model according to the chosen \code{criterion}.
}
\description{
selectRHLP implements the model selection procedure.
}
\details{
selectRHLP implements the model selection. This function runs every
RHLP model by varying the number of regimes \code{K} from \code{Kmin} to \code{Kmax} and
the order of the polynomial regression \code{p} from \code{pmin} to \code{pmax}. The model
having the highest value of the chosen selection criterion is then
selected.
}
\seealso{
\link{ModelRHLP}
}
