% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-univtoydataset.R
\docType{data}
\name{univtoydataset}
\alias{univtoydataset}
\title{A simulated non-stationary time series with regime changes.}
\format{A data frame with 670 rows and 2 columns:
\describe{
\item{x}{The covariate variable which is the time in this time-series case.}
\item{y}{The time series. The latter has been generated as follows:
\itemize{
\item  First regime: 100 values of standard Normally distributed random numbers.
\item Second regime: 120 values of Normally distributed random numbers
with mean 7 and unit variance.
\item Third regime: 200 values of Normally distributed random numbers
with mean 4  and unit variance.
\item Fourth regime: 100 values of Normally distributed random numbers
with mean -2 and unit variance.
\item Fifth regime: 150 values of Normally distributed random numbers
with mean 3.5  and unit variance.
}
}
}}
\usage{
univtoydataset
}
\description{
A simulated non-stationary time series with regime changes.
This time series is used for illustration.
}
\keyword{datasets}
